`qml.simplify` can reduce number of rotation parameters (#2905)
* move rotation simplify into op method

* add additional example
