+++ pennylane/ops/functions/simplify.py
@@ -22,7 +22,8 @@ from pennylane.operation import Operator
 
 @qml.op_transform
 def simplify(op: Operator):
-    """Reduces the depth of nested operators.
+    """Simplifies the operator by reducing arithmetic depth or number of rotation
+    parameters.
 
 
     Args:
@@ -47,6 +48,20 @@ def simplify(op: Operator):
     (Adjoint(RX)(0.54, wires=[0]),
     Adjoint(PauliX)(wires=[0]),
     Adjoint(PauliZ)(wires=[1]))
+
+    This function also can simplify the number of rotation gate parameters:
+
+    >>> qml.simplify(qml.Rot(np.pi / 2, 0.1, -np.pi / 2, wires=0))
+    RX(0.1, wires=[0])
+
+    Both types of simplification occur together:
+
+    >>> op = qml.adjoint(qml.U2(-np.pi/2, np.pi/2, wires=0) + qml.PauliX(0))
+    >>> op
+    Adjoint(Sum)([-1.5707963267948966, 1.5707963267948966], [], wires=[0])
+    >>> qml.simplify(op)
+    Adjoint(RX)(1.5707963267948966, wires=[0]) + Adjoint(PauliX)(wires=[0])
+
     """
     if qml.queuing.QueuingContext.recording():
         with qml.tape.stop_recording():
